requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
watchdog>=6.0.0
pymupdf>=1.25.0
pymupdf4llm>=0.0.17
chromadb>=0.6.0
networkx>=3.4.0
rank_bm25>=0.2.2
python-frontmatter>=1.1.0
pillow>=11.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
python-dotenv>=1.0.0
mcp>=1.0.0
nh3>=0.2.0
